Creating useListItemManager manage the item click method. using walkme's SDK for list items
