--- src/app/hooks/useLink.tsx
@@ -1,20 +1,0 @@
-import React from "react";
-import { useHistory } from "react-router-dom";
-
-export default function useLink() {
-  const routeHistory = useHistory();
-
-  const handleLinkClick = (link: string) => {
-    console.log("handleLinkClick ", link);
-    const isExternalLink =
-      link.includes("http://") || link.includes("https://");
-
-    if (isExternalLink) {
-      window.open(link, "_blank");
-    } else {
-      routeHistory.push(link);
-    }
-  };
-
-  return { handleLinkClick };
-}
